Move some generic data structures from under `single_disk_farm` into `farm` module, no other code changes
